+++ migrations/20250323190249_versions_last_update.sql
@@ -1,0 +1,3 @@
+-- Add migration script here
+ALTER TABLE programs ADD current_version_last_updated DATETIME NOT NULL DEFAULT '1970-01-01 00:00:00';
+ALTER TABLE programs ADD latest_version_last_updated DATETIME NOT NULL DEFAULT '1970-01-01 00:00:00';