+++ changelog.md
@@ -3,6 +3,7 @@
 ## v1.1.0
 
 - Last time an update check was performed is now displayed when command 'list-programs' is run
+- Current version last updated and latest version last updated times are now saved and are displayed in 'list-programs'
 
 ## v1.0.0